emulator/pyproject.toml
[project]
name = "robops-emulator"
version = "0.1.0"
description = "RobOps Platform - Robot Emulator"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "awsiotsdk>=1.22.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]
